feat: add mock lifecycle event listener into new file
